% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot fatalities on a map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the number of the state to plot (coerced to integer).}

\item{year}{an integer representing the year to plot .}
}
\value{
None
}
\description{
\code{fars_map_state} plots fatalities on a map for a state in a given year
}
\examples{
\dontrun{fars_map_state(2, 2015)}

}
