% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a fars file and return a data frame tbl}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the name of the file to read (character)}
}
\value{
a data frame tbl
}
\description{
A function to read comma separated values fars files.
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")
fars_read(make_filename(2013))}


}
