% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{A helper function for \code{\link{fars_summarize_years}}}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of integer specifying the years to extract}
}
\value{
a list of data frames
}
\description{
Extracts month/year couples from each year in \code{years} and stores it in a list. Return
an error if year in \code{years} does not exist in the data.
}
\examples{
make_filename(2013)

}
