% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Number of fatalities by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of integer specifying the years to tabulate.}
}
\value{
a data frame
}
\description{
Returns a month/year table with the number of fatalities for each month of each year.
}
\examples{
fars_summarize_years(2013:2015)

}
