% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Fars file name for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year of the fars data file (coerced to integer)}
}
\value{
a string
}
\description{
\code{make_filename()} takes a year number and returns the name of the file containing
the fars data for this year.
}
\examples{
make_filename(2013)

}
